requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-pptx==0.6.23
jinja2==3.1.3
aiofiles==23.2.1
python-dotenv==1.0.1
requests==2.31.0
unoserver==2.0.1
httpx[http2]

# Google Drive API
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
